requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
ansicolors==1.1.8
anthropic==0.75.0
anyio==4.12.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.3.0
build==1.3.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.1.8
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
crewai==1.7.2
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
entrypoints==0.4
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.1
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
instructor==1.12.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.3
json5==0.10.0
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==1.2.0
langchain-anthropic==1.3.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-google-genai==4.1.2
langchain-ollama==1.0.1
langchain-text-splitters==1.1.0
langgraph==1.0.5
sqlglot==25.0.0
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.1
lark==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mcp==1.16.0
mdurl==0.1.2
mistune==3.2.0
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.14.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.10.0
notebook==7.5.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2
openai==1.83.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
papermill==2.6.0
parso==0.8.5
pdfminer.six==20251107
pdfplumber==0.11.8
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.5.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.9.0
pyparsing==3.3.1
pypdfium2==5.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-asyncio==1.3.0
testcontainers[postgres]==4.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-magic==0.4.27
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2024.9.11
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
sse-starlette==3.2.0
stack-data==0.6.3
starlette==0.50.0
fastapi==0.128.0
slowapi==0.1.9
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.0.2
tomli_w==1.1.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typer==0.21.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.3.0
uuid_utils==0.12.0
uv==0.9.18
uvicorn==0.40.0
uvloop==0.22.1
virtualenv==20.35.4
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.15
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
litellm==1.40.8
xgboost==2.0.3
asyncpg==0.29.0
alembic==1.13.1

python-socketio==5.11.0
tiktoken==0.5.2



psycopg2-binary==2.9.9
redis==7.1.0
celery==5.6.2
werkzeug==3.1.5
pyarrow==14.0.1
prophet==1.1.5
lightgbm==4.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Cloud Data Warehouse Drivers (optional - install manually as needed)
# sqlalchemy-redshift is incompatible with SQLAlchemy 2.x; use redshift-connector instead.
# pip install sqlalchemy-bigquery==1.12.0 snowflake-sqlalchemy==1.7.2 redshift-connector

# Cloud Storage Connectors (optional - install as needed)
boto3==1.35.0
google-cloud-storage==2.18.0
azure-storage-blob==12.23.0

# REST API Data Source
jmespath==1.0.1

# Report Export
weasyprint==62.3
python-pptx==1.0.2
markdown==3.7
